requirements.txt
pillow
tensorflow
numpy
mouse
matplotlib